Now admin can't delete a departament if there exists a doctor in it, cause it will raise an error in the database level(doctors table has a foreign key to departaments table.

+++ app/Http/Controllers/AdminController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Departament;
+use App\Models\Doctor;
 
 class AdminController extends Controller
 {
@@ -42,7 +43,13 @@ class AdminController extends Controller
     public function deleteDepartament($id)
     {
         $departament = Departament::findOrFail($id);
-        $departament->delete();
-        return redirect()->route('show-departaments')->with('test', 'Departament deleted successfully.');
+        $doctor = Doctor::where('departament_id', $id)->first();
+        if ($doctor) {
+            $message = 'Departament can\'t be deleted since a doctor is in that departament. His name is '.$doctor->first_name;
+            return redirect()->route('show-departaments')->with('test', $message);
+        } else {
+            $departament->delete();
+            return redirect()->route('show-departaments')->with('test', 'Departament deleted successfully.');
+        }
     }
 }
